feat: implement Figma import functionality with separate dialogs for .fig and .deck files
- Added ConfirmFigImportCard component to display import confirmation details.
- Created ImportFromFigmaApiDialog for importing Figma content via REST API.
- Developed ImportFromFigmaFileDialog for importing .fig files with a dropzone.
- Introduced ImportFromFigmaSlidesDialog for importing .deck files.
- Removed legacy ImportFromFigmaDialog and refactored import logic into dedicated hooks and components.
- Added useFigFileImport hook to manage file import state and parsing.
- Updated index.ts to export new import components and types.
