+++ editor/grida-canvas-react-starter-kit/starterkit-import/index.ts
@@ -1,2 +1,8 @@
 export { ImportFromGridaDialog } from "./from-grida";
-export { ImportFromFigmaDialog, type FetchNodeResult } from "./from-figma";
+export { ImportFromFigmaFileDialog } from "./from-figma-file";
+export { ImportFromFigmaSlidesDialog } from "./from-figma-slides";
+export {
+  ImportFromFigmaApiDialog,
+  type FetchNodeResult,
+} from "./from-figma-api";
+export type { FigFileImportResult } from "./use-fig-file-import";
